Refactor: move FieldTreeTable/Model to own classes
