+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/normalizedcache/NormalizedCacheToolWindowFactory.kt
@@ -3,15 +3,9 @@ package com.apollographql.ijplugin.normalizedcache
 import com.apollographql.ijplugin.ApolloBundle
 import com.apollographql.ijplugin.apollodebugserver.ApolloDebugClient
 import com.apollographql.ijplugin.apollodebugserver.normalizedCacheSimpleName
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.BooleanValue
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.CompositeValue
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.ListValue
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.Null
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.NumberValue
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.Reference
-import com.apollographql.ijplugin.normalizedcache.NormalizedCache.FieldValue.StringValue
 import com.apollographql.ijplugin.normalizedcache.provider.ApolloDebugNormalizedCacheProvider
 import com.apollographql.ijplugin.normalizedcache.provider.DatabaseNormalizedCacheProvider
+import com.apollographql.ijplugin.normalizedcache.ui.FieldTreeTable
 import com.apollographql.ijplugin.normalizedcache.ui.RecordSearchTextField
 import com.apollographql.ijplugin.normalizedcache.ui.RecordTable
 import com.apollographql.ijplugin.telemetry.TelemetryEvent
@@ -20,8 +14,6 @@ import com.apollographql.ijplugin.util.logw
 import com.apollographql.ijplugin.util.showNotification
 import com.intellij.icons.AllIcons
 import com.intellij.ide.CommonActionsManager
-import com.intellij.ide.DefaultTreeExpander
-import com.intellij.ide.TreeExpander
 import com.intellij.ide.dnd.FileCopyPasteUtil
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.Disposable
@@ -35,41 +27,26 @@ import com.intellij.openapi.actionSystem.ex.ActionUtil
 import com.intellij.openapi.application.invokeLater
 import com.intellij.openapi.fileChooser.FileChooser
 import com.intellij.openapi.fileChooser.FileChooserDescriptor
-import com.intellij.openapi.ide.CopyPasteManager
 import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.progress.Task
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.DumbAwareAction
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.ui.JBMenuItem
-import com.intellij.openapi.ui.JBPopupMenu
 import com.intellij.openapi.ui.SimpleToolWindowPanel
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.openapi.wm.ex.ToolWindowEx
 import com.intellij.openapi.wm.ex.ToolWindowManagerListener
-import com.intellij.ui.ColoredTableCellRenderer
 import com.intellij.ui.OnePixelSplitter
 import com.intellij.ui.ScrollPaneFactory
 import com.intellij.ui.SimpleTextAttributes
 import com.intellij.ui.components.JBPanelWithEmptyText
-import com.intellij.ui.components.JBTreeTable
 import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.content.ContentManager
-import com.intellij.ui.table.JBTable
-import com.intellij.ui.treeStructure.treetable.ListTreeTableModel
-import com.intellij.ui.treeStructure.treetable.TreeTableModel
-import com.intellij.util.ui.ColumnInfo
 import com.intellij.util.ui.JBUI
-import com.intellij.util.ui.UIUtil
 import kotlinx.coroutines.runBlocking
 import org.sqlite.SQLiteException
 import java.awt.BorderLayout
-import java.awt.Color
-import java.awt.Component
-import java.awt.Cursor
-import java.awt.Point
-import java.awt.datatransfer.StringSelection
 import java.awt.dnd.DnDConstants
 import java.awt.dnd.DropTarget
 import java.awt.dnd.DropTargetAdapter
@@ -77,18 +54,10 @@ import java.awt.dnd.DropTargetDragEvent
 import java.awt.dnd.DropTargetDropEvent
 import java.awt.dnd.DropTargetEvent
 import java.awt.event.KeyEvent
-import java.awt.event.MouseAdapter
-import java.awt.event.MouseEvent
 import java.io.File
 import javax.swing.JComponent
 import javax.swing.JPanel
-import javax.swing.JTable
 import javax.swing.SwingUtilities
-import javax.swing.event.PopupMenuEvent
-import javax.swing.event.PopupMenuListener
-import javax.swing.tree.DefaultMutableTreeNode
-import javax.swing.tree.TreePath
-
 
 class NormalizedCacheToolWindowFactory : ToolWindowFactory, DumbAware, Disposable {
   override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
@@ -137,11 +106,9 @@ class NormalizedCacheWindowPanel(
   private lateinit var normalizedCache: NormalizedCache
 
   private lateinit var recordTable: RecordTable
-
-  private lateinit var fieldTreeTableModel: ListTreeTableModel
   private lateinit var recordSearchTextField: RecordSearchTextField
 
-  private lateinit var fieldTreeExpander: TreeExpander
+  private lateinit var fieldTreeTable: FieldTreeTable
 
   private val history = History<NormalizedCache.Record>()
   private var updateHistory = true
@@ -258,10 +225,10 @@ class NormalizedCacheWindowPanel(
         override fun getActionUpdateThread() = ActionUpdateThread.BGT
       })
       addSeparator()
-      add(CommonActionsManager.getInstance().createExpandAllAction(fieldTreeExpander, this@NormalizedCacheWindowPanel).apply {
+      add(CommonActionsManager.getInstance().createExpandAllAction(fieldTreeTable.treeExpander, this@NormalizedCacheWindowPanel).apply {
         getTemplatePresentation().setDescription(ApolloBundle.message("normalizedCacheViewer.toolbar.expandAll"))
       })
-      add(CommonActionsManager.getInstance().createCollapseAllAction(fieldTreeExpander, this@NormalizedCacheWindowPanel).apply {
+      add(CommonActionsManager.getInstance().createCollapseAllAction(fieldTreeTable.treeExpander, this@NormalizedCacheWindowPanel).apply {
         getTemplatePresentation().setDescription(ApolloBundle.message("normalizedCacheViewer.toolbar.collapseAll"))
       })
       addSeparator()
@@ -291,13 +258,16 @@ class NormalizedCacheWindowPanel(
   }
 
   private fun createNormalizedCacheContent(): JComponent {
-    val fieldTreeTable = createFieldTreeTable()
+    fieldTreeTable = FieldTreeTable(::selectRecord)
     val recordTableWithFilter = createRecordTableWithFilter()
     val splitter = OnePixelSplitter(false, .25F).apply {
       firstComponent = recordTableWithFilter
       secondComponent = fieldTreeTable
       setResizeEnabled(true)
       splitterProportionKey = "${NormalizedCacheToolWindowFactory::class.java}.splitterProportionKey"
+    }
+    SwingUtilities.invokeLater {
+      recordTable.requestFocusInWindow()
     }
     return splitter
   }
@@ -308,7 +278,7 @@ class NormalizedCacheWindowPanel(
         if (selectedRow == -1) return@addListSelectionListener
         val selectedRowAfterSort = convertRowIndexToModel(selectedRow)
         model.getRecordAt(selectedRowAfterSort)?.let { record ->
-          fieldTreeTableModel.setRoot(getRootNodeForRecord(record))
+          fieldTreeTable.setRecord(record)
           if (!updateHistory) {
             updateHistory = true
           } else {
@@ -330,154 +300,6 @@ class NormalizedCacheWindowPanel(
     return tableWithFilter
   }
 
-  @Suppress("UnstableApiUsage")
-  private fun createFieldTreeTable(): JComponent {
-    fieldTreeTableModel = ListTreeTableModel(
-        DefaultMutableTreeNode(),
-        arrayOf(
-            object : ColumnInfo<Unit, Unit>(ApolloBundle.message("normalizedCacheViewer.fields.column.key")) {
-              override fun getColumnClass() = TreeTableModel::class.java
-              override fun valueOf(item: Unit) = Unit
-            },
-            object : ColumnInfo<NormalizedCacheFieldTreeNode, NormalizedCache.Field>(ApolloBundle.message("normalizedCacheViewer.fields.column.value")) {
-              override fun getColumnClass() = NormalizedCache.Field::class.java
-              override fun valueOf(item: NormalizedCacheFieldTreeNode) = item.field
-            },
-        ),
-    )
-
-    val treeTable = object : JBTreeTable(fieldTreeTableModel) {
-      override fun getPathBackground(path: TreePath, row: Int): Color? {
-        return if (row % 2 == 0) {
-          UIUtil.getDecoratedRowColor()
-        } else {
-          null
-        }
-      }
-    }.apply {
-      columnProportion = .8F
-      setDefaultRenderer(NormalizedCache.Field::class.java, object : ColoredTableCellRenderer() {
-        override fun customizeCellRenderer(table: JTable, value: Any?, selected: Boolean, hasFocus: Boolean, row: Int, column: Int) {
-          value as NormalizedCache.Field
-          when (val v = value.value) {
-            is StringValue -> append("\"${v.value}\"")
-            is NumberValue -> append(v.value.toString())
-            is BooleanValue -> append(v.value.toString())
-            is ListValue -> append(when (val size = v.value.size) {
-              0 -> ApolloBundle.message("normalizedCacheViewer.fields.list.empty")
-              1 -> ApolloBundle.message("normalizedCacheViewer.fields.list.single")
-              else -> ApolloBundle.message("normalizedCacheViewer.fields.list.multiple", size)
-            }, SimpleTextAttributes.GRAY_ITALIC_ATTRIBUTES)
-
-            is CompositeValue -> append("{...}", SimpleTextAttributes.GRAY_ITALIC_ATTRIBUTES)
-            Null -> append("null")
-            is Reference -> {
-              append("→ ", SimpleTextAttributes.GRAY_ITALIC_ATTRIBUTES)
-              append(v.key, SimpleTextAttributes.LINK_PLAIN_ATTRIBUTES)
-            }
-          }
-        }
-      })
-
-      // Handle reference clicks and cursor changes
-      val mouseAdapter = object : MouseAdapter() {
-        override fun mouseClicked(e: MouseEvent) {
-          if (e.button != MouseEvent.BUTTON1) return
-          table.cursor = Cursor(Cursor.DEFAULT_CURSOR)
-          val field = table.getFieldAtPoint(e) ?: return
-          if (field.value is Reference) {
-            selectRecord(field.value.key)
-          }
-        }
-
-        override fun mouseMoved(e: MouseEvent) {
-          table.cursor = Cursor(
-              if (table.getFieldAtPoint(e)?.value is Reference) {
-                Cursor.HAND_CURSOR
-              } else {
-                Cursor.DEFAULT_CURSOR
-              }
-          )
-        }
-
-        override fun mouseExited(e: MouseEvent) {
-          table.cursor = Cursor(Cursor.DEFAULT_CURSOR)
-        }
-      }
-      table.addMouseListener(mouseAdapter)
-      table.addMouseMotionListener(mouseAdapter)
-
-      table.isStriped = true
-
-      installPopupMenu(table)
-
-      fieldTreeExpander = DefaultTreeExpander { tree }
-    }
-    return treeTable
-  }
-
-  private fun installPopupMenu(table: JBTable) {
-    val popupMenu = object : JBPopupMenu() {
-      override fun show(invoker: Component, x: Int, y: Int) {
-        when (table.getFieldAtPoint(x, y)?.value) {
-          is StringValue,
-          is NumberValue,
-          is BooleanValue,
-          is Reference,
-          Null,
-          -> {
-            super.show(invoker, x, y)
-          }
-
-          else -> {}
-        }
-      }
-    }
-    popupMenu.add(JBMenuItem(ApolloBundle.message("normalizedCacheViewer.fields.popupMenu.copyValue")).apply {
-      addActionListener {
-        val field = table.getValueAt(table.selectedRow, table.selectedColumn) as NormalizedCache.Field
-        val valueStr = when (val value = field.value) {
-          is StringValue -> value.value
-          is NumberValue -> value.value.toString()
-          is BooleanValue -> value.value.toString()
-          is Reference -> value.key
-          Null -> "null"
-          else -> return@addActionListener
-        }
-        CopyPasteManager.getInstance().setContents(StringSelection(valueStr))
-      }
-    })
-
-    // Select the row under the popup menu
-    popupMenu.addPopupMenuListener(object : PopupMenuListener {
-      override fun popupMenuWillBecomeVisible(e: PopupMenuEvent) {
-        invokeLater {
-          val row = table.rowAtPoint(SwingUtilities.convertPoint(popupMenu, Point(0, 0), table))
-          val column = table.columnAtPoint(SwingUtilities.convertPoint(popupMenu, Point(0, 0), table))
-          table.setRowSelectionInterval(row, row)
-          table.setColumnSelectionInterval(column, column)
-        }
-      }
-
-      override fun popupMenuWillBecomeInvisible(e: PopupMenuEvent) {}
-
-      override fun popupMenuCanceled(e: PopupMenuEvent) {}
-    })
-    table.componentPopupMenu = popupMenu
-  }
-
-  private fun JBTable.getFieldAtPoint(x: Int, y: Int): NormalizedCache.Field? {
-    val point = Point(x, y)
-    val row = rowAtPoint(point)
-    // Add 1 to account for the tree column
-    val column = columnAtPoint(point) + 1
-    return model.getValueAt(row, column) as NormalizedCache.Field?
-  }
-
-  private fun JBTable.getFieldAtPoint(e: MouseEvent): NormalizedCache.Field? {
-    return getFieldAtPoint(e.x, e.y)
-  }
-
   private fun selectRecord(key: String) {
     if (!recordTable.model.isRecordShowing(key)) {
       // Filtered list doesn't contain the record, so clear the filter
@@ -485,22 +307,6 @@ class NormalizedCacheWindowPanel(
     }
     recordTable.selectRecord(key)
     recordTable.requestFocusInWindow()
-  }
-
-  private fun getRootNodeForRecord(record: NormalizedCache.Record) = DefaultMutableTreeNode().apply {
-    addFields(record.fields)
-  }
-
-  private fun DefaultMutableTreeNode.addFields(fields: List<NormalizedCache.Field>) {
-    for (field in fields) {
-      val childNode = NormalizedCacheFieldTreeNode(field)
-      add(childNode)
-      when (val value = field.value) {
-        is ListValue -> childNode.addFields(value.value.mapIndexed { i, v -> NormalizedCache.Field(i.toString(), v) })
-        is CompositeValue -> childNode.addFields(value.value)
-        else -> {}
-      }
-    }
   }
 
   private fun pickFile() {
@@ -607,12 +413,6 @@ class NormalizedCacheWindowPanel(
         }
       }
     }.queue()
-  }
-
-  private class NormalizedCacheFieldTreeNode(val field: NormalizedCache.Field) : DefaultMutableTreeNode() {
-    init {
-      userObject = field.name
-    }
   }
 
   override fun dispose() {
